Remove --whole-archive and call init functions for pyversion and pyodide_pre
Object files in an archive are only included if at least one file
is needed from them. We need _pyodide_core.o directly because it
defines `PyInit__pyodide_core` which we add to the Inittab. Each
other object file must be hereditarily required by `_pyodide_core.o`.
Most object files are needed because of the TRY_INIT macros in
`PyInit__pyodide_core` but `pyversion.o` and `pyodide_pre.o` were not.
This fixes it.